refactor: redo headless mode as a runtime, not link, option
The link approach was very clean in a C-ish way, but it doesn't reflect
the most important use case of a mode like this: running tests on a
build that is otherwise playable. This commit refactors the libheadless
code back into libunix.cc in a brute-force way.

In addition, I discovered that tiles already had a proto-implementation
of this idea in the form of crawl_state.tiles_disabled, aimed at mapstat
and objstat. Actually getting tiles headless mode to work for
scripts/tests is a fairly tall order, but I've somewhat merged the API
and CLI for these two things without going that far. The only build that
-headless is fully unavailable in should be windows console. (However,
windows tiles not tested.)

This commit also forces headless mode for -script and -test for builds
that can do that.

Still not working: webtiles headless processes, running rc scripts in
headless mode.